server/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
anyio==3.5.0; python_full_version >= '3.6.2'
asgiref==3.4.1; python_version >= '3.6'
beautifulsoup4==4.10.0; python_version >= '3.1'
blis==0.7.5
boto3==1.20.37; python_version >= '3.6'
botocore==1.23.37; python_version >= '3.6'
bottleneck==1.3.2
catalogue==2.0.6; python_version >= '3.6'
certifi==2021.10.8
charset-normalizer==2.0.10; python_version >= '3'
click==8.0.3; python_version >= '3.6'
cycler==0.11.0; python_version >= '3.6'
cymem==2.0.6
fastai==1.0.60
fastapi==0.71.0
fastprogress==1.0.0; python_version >= '3.6'
filelock==3.4.2; python_version >= '3.7'
fonttools==4.28.5; python_version >= '3.7'
h11==0.12.0; python_version >= '3.6'
httptools==0.3.0
idna==3.3; python_version >= '3.5'
importlib-metadata==4.10.0; python_version < '3.8'
jinja2==3.0.3; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
joblib==1.1.0; python_version >= '3.6'
kiwisolver==1.3.2; python_version >= '3.7'
langcodes==3.3.0; python_version >= '3.6'
markupsafe==2.0.1; python_version >= '3.6'
matplotlib==3.5.1; python_version >= '3.7'
murmurhash==1.0.6
numexpr==2.8.1
numpy==1.21.5; python_version < '3.11' and python_version >= '3.7'
nvidia-ml-py3==7.352.0
oauthlib==3.1.1; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pandas==1.3.5; python_full_version >= '3.7.1'
pathy==0.6.1; python_version >= '3.6'
pillow==9.0.0; python_version >= '3.7'
preshed==3.0.6
pydantic==1.8.2; python_full_version >= '3.6.1'
pynvx==1.0.0; platform_system == 'Darwin'
pyparsing==3.0.6; python_version >= '3.6'
pysocks==1.7.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.19.2
pytz==2021.3
pyyaml==6.0; python_version >= '3.6'
regex==2021.11.10
requests-oauthlib==1.3.0
requests[socks]==2.27.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
s3transfer==0.5.0; python_version >= '3.6'
sacremoses==0.0.47
scipy==1.7.3; python_version < '3.11' and python_version >= '3.7'
sentencepiece==0.1.96
setuptools==60.5.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smart-open==5.2.1; python_version >= '3.6' and python_version < '4'
sniffio==1.2.0; python_version >= '3.5'
soupsieve==2.3.1; python_version >= '3.6'
spacy-legacy==3.0.8; python_version >= '3.6'
spacy-loggers==1.0.1; python_version >= '3.6'
spacy==3.2.1; python_version >= '3.6'
srsly==2.4.2; python_version >= '3.6'
starlette==0.17.1; python_version >= '3.6'
thinc==8.0.13; python_version >= '3.6'
tokenizers==0.5.2
torch==1.10.1; python_full_version >= '3.6.2'
torchvision==0.11.2
tqdm==4.62.3; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
transformers==2.7.0
tweepy==3.10.0
typer==0.4.0; python_version >= '3.6'
typing-extensions==3.10.0.2; python_version < '3.8'
urllib3==1.26.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
uvicorn[standard]==0.17.0
uvloop==0.16.0
wasabi==0.9.0
watchgod==0.7
websockets==10.1
zipp==3.7.0; python_version < '3.8'
